% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_dist.R
\docType{data}
\name{cov_dist}
\alias{cov_dist}
\title{Coverage Probability by Distribution Plot}
\format{A ggplot object.}
\usage{
cov_dist
}
\description{
A ggplot of coverae probability by distribution. Facets by sample size and method.
}
\keyword{datasets}
