% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff-est-bland.R
\name{eff_est_bland}
\alias{eff_est_bland}
\title{Bland effect estimator}
\usage{
eff_est_bland(a, q.1, m, q.3, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{q.1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q.3}{Third quartile of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Calculate effects' estimates for mean and sd using Bland's method.
}
