% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff-est-hozo.R
\name{eff_est_hozo}
\alias{eff_est_hozo}
\title{Hozo effect estimator}
\usage{
eff_est_hozo(a, m, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{m}{Median of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Calculate effects' estimates for mean and sd using Hozo's method. This formula
was taken from Wan's paper, where Hozo's method is summarised, equation (6).
}
