% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff-est-wan-c1.R
\name{eff_est_wan_c1}
\alias{eff_est_wan_c1}
\title{WanC1 effect estimator}
\usage{
eff_est_wan_c1(a, m, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{m}{Median of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Calculate effects' estimates for mean and sd using WanC1's method.
}
