% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff-est-wan-c2.R
\name{eff_est_wan_c2}
\alias{eff_est_wan_c2}
\title{WanC2 effect estimator}
\usage{
eff_est_wan_c2(a, q.1, m, q.3, b, n)
}
\arguments{
\item{a}{Minimum value of sample.}

\item{q.1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q.3}{Third quartile of sample.}

\item{b}{Maximum value of sample.}

\item{n}{Sample size.}
}
\description{
Calculate effects' estimates for mean and sd using WanC2's method.
}
