% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff-est-wan-c3.R
\name{eff_est_wan_c3}
\alias{eff_est_wan_c3}
\title{WanC3 effect estimator}
\usage{
eff_est_wan_c3(q.1, m, q.3, n)
}
\arguments{
\item{q.1}{First quartile of sample.}

\item{m}{Median of sample.}

\item{q.3}{Third quartile of sample.}

\item{n}{Sample size.}
}
\description{
Calculate effects' estimates for mean and sd using WanC3's method.
}
