% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo_12.R
\name{hozo_16}
\alias{hozo_16}
\title{Hozo Equation 16}
\usage{
hozo_16(min, median, max)
}
\arguments{
\item{min}{Minimum of the sample.}

\item{median}{Median of the sample.}

\item{max}{Maximum of the sample.}
}
\description{
Equation (16) from Hozo's paper. Estimates an approximation of the sample
standard deviation from some summary statistics.
}
