% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo_5.R
\name{hozo_5}
\alias{hozo_5}
\title{Hozo Equation 5}
\usage{
hozo_5(min, median, max)
}
\arguments{
\item{min}{Minimum of the sample.}

\item{median}{Median of the sample.}

\item{max}{Maximum of the sample.}
}
\description{
Equation (5) from Hozo's paper. Estimates an approximation of the sample
mean from some summary statistics.
}
