% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo-mean-4.R
\name{hozo_mean_4}
\alias{hozo_mean_4}
\title{Hozo (4)}
\usage{
hozo_mean_4(a, b, m, n)
}
\arguments{
\item{a}{Sample minimum.}

\item{b}{Sample maximum.}

\item{m}{Sample median.}

\item{n}{Sample size.}
}
\description{
Hozo et al.'s estimator for the sample mean from formula (4).
}
