% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo-mean-5.R
\name{hozo_mean_5}
\alias{hozo_mean_5}
\title{Hozo (5)}
\usage{
hozo_mean_5(a, b, m)
}
\arguments{
\item{a}{Sample minimum.}

\item{b}{Sample maximum.}

\item{m}{Sample median.}
}
\description{
Hozo et al.'s estimator for the sample mean from formula (5). Note that this
one does not depend on \eqn{n}.
}
