% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo-sd-12.R
\name{hozo_sd_12}
\alias{hozo_sd_12}
\title{Hozo (12)}
\usage{
hozo_sd_12(a, b, m, n)
}
\arguments{
\item{a}{Sample minimum.}

\item{b}{Sample maximum.}

\item{m}{Sample median.}

\item{n}{Sample size.}
}
\description{
Hozo et al.'s estimator for the sample standard deviation from formula (12).
}
