% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hozo-sd-16.R
\name{hozo_sd_16}
\alias{hozo_sd_16}
\title{Hozo (16)}
\usage{
hozo_sd_16(a, b, m)
}
\arguments{
\item{a}{Sample minimum.}

\item{b}{Sample maximum.}

\item{m}{Sample median.}
}
\description{
Hozo et al.'s estimator for the sample standard deviation from formula (12).
}
